% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pba2024.R
\name{load_pba2024}
\alias{load_pba2024}
\title{load_pba2024()}
\usage{
load_pba2024()
}
\value{
(the tibble)
}
\description{
Loads the PBA Tour 2024 player analysis from an excel file. Note that this data comes
from the year 2023.
}
