% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pbaseniorsusopen2024.R
\name{load_pbaseniorusopen2024}
\alias{load_pbaseniorusopen2024}
\title{load_pbaseniorusopen2024()}
\usage{
load_pbaseniorusopen2024()
}
\value{
(the tibble)
}
\description{
Loads the PBA Senior US Open 2024 final standing results from an excel file.
}
