% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_summary.R
\name{player_summary}
\alias{player_summary}
\title{player_summary()}
\usage{
player_summary(dataset, player)
}
\arguments{
\item{dataset}{The dataset used to get information.}

\item{player}{The player to be searched for.}
}
\value{
Summarized player stats.
}
\description{
Summarizes the chosen player's stats using the chosen dataset.
}
