% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot_average.R
\name{scatterplot_average}
\alias{scatterplot_average}
\title{scatterplot_average()}
\usage{
scatterplot_average(
  dataset,
  y_axis = "Earnings",
  linear_regression = FALSE,
  lregression_method = "lm"
)
}
\arguments{
\item{dataset}{Dataset to be used to get info.}

\item{y_axis}{The y axis values to be plotted.}

\item{linear_regression}{Adds a linear regression line. Set to FALSE on default.}

\item{lregression_method}{Method used to do the linear regression line. Set to lm on default}
}
\description{
Scatterplots the average of the player against a chosen factor (earnings by default)
}
